handle an empty cache directory properly
now, the cache directory has to be specified, and there is a reasonable
default, but it doesn't have to be any more than specified. if the user
says "/" for the cache directory, tarsnap may protest but the GUI accepts
that. possibly no ideal but I like it better than checking strictly.